feat(api): add todo testing
